adjustments on services

+++ src/main/java/br/com/usp/labis/controller/EnrichmentAnalysisController.java
@@ -37,7 +37,7 @@ public class EnrichmentAnalysisController {
 	
 	@CrossOrigin
 	@PostMapping
-	public String processEnrichmentAnalysis(@RequestParam("file") MultipartFile file,
+	public ResponseEntity<String> processEnrichmentAnalysis(@RequestParam("file") MultipartFile file,
 			@RequestParam("taxonId") Integer taxonId, @RequestParam("minProteins") Integer minProteins,
 			@RequestParam("toleranceFactor") Double toleranceFactor,
 			@RequestParam("nullDistributions") Integer nullDistributions, @RequestParam("pvalue") Double pvalue) {
@@ -45,16 +45,22 @@ public class EnrichmentAnalysisController {
 		String resultFilePath = enrichmentAnalysisService.processEnrichmentAnalysis(file, taxonId, minProteins,
 				toleranceFactor, nullDistributions, pvalue);
 		
-		return resultFilePath;
-
+		HttpHeaders headers = new HttpHeaders();
+	    headers.add("Content-Type", "application/json; charset=utf-8");
+		
+		System.out.println("resultFilePath: " + resultFilePath);
+		
+		return new ResponseEntity<>(resultFilePath,  headers , HttpStatus.OK);
 	}
 
 
 	@CrossOrigin
-	@GetMapping("/download/{fileName}")
-	public ResponseEntity<Resource> downloadFile(@PathParam("fileName")  String fileName, HttpServletResponse response) {
-		String filePath = "C:" + File.separator + "uploaded_file" + File.separator + fileName;
-		File fileOutput = outputFileService.getFileByName(fileName);
+	@GetMapping("/download/{filepath}")
+	public ResponseEntity<Resource> downloadFile(@PathParam("filepath")  String filepath, HttpServletResponse response) {
+
+		System.out.println("Downloading: " + filepath);
+
+		File fileOutput = outputFileService.getFileByName(filepath);
 
 		HttpHeaders headers = new HttpHeaders();
 		headers.setContentType(MediaType.APPLICATION_OCTET_STREAM);
@@ -66,9 +72,10 @@ public class EnrichmentAnalysisController {
 	
 	@CrossOrigin
 	@GetMapping("/delete/{fileName}")
-	public void deleteFile(@PathParam("fileName") String fileName) {
-		String filePath = "C:" + File.separator + "uploaded_file" + File.separator + fileName;
-		File file = outputFileService.getFileByName(fileName);
-		uploadFileService.removeUploadedFile(file);
+	public void deleteFile(@PathParam("fileName") String filepath) {
+		File file = outputFileService.getFileByName(filepath);
+		if(file != null) {
+			uploadFileService.removeUploadedFile(file);
+		}
 	}
 }
